requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
requests==2.32.3
pandas==2.2.2
pydantic==2.8.2
pydantic-settings==2.3.4
python-telegram-bot==21.4
APScheduler==3.10.4
yfinance==0.2.40
